feat: day 3 - alembic migrations setup

+++ backend/app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column,Integer,String,Float,ForeignKey,UniqueConstraint
+from sqlalchemy import Column,Integer,String,Float,ForeignKey,UniqueConstraint,Numeric
 from .database import Base
 from sqlalchemy.sql.sqltypes import TIMESTAMP
 from sqlalchemy.sql.expression import text
@@ -13,7 +13,7 @@ class Expense(Base):
     __tablename__="expenses"
     id=Column(Integer,primary_key=True,index=True)
     title=Column(String,nullable=False,index=True)
-    amount=Column(Float,nullable=False,index=True)
+    amount=Column(Numeric(10,2),nullable=False,index=True)
     description=Column(String,nullable=False)
     created_at=Column(TIMESTAMP(timezone=True),nullable=False,server_default=text('now()'))
     owner_id=Column(Integer,ForeignKey("users.id"),nullable=False)
@@ -22,7 +22,7 @@ class Expense(Base):
 class Budget(Base):
     __tablename__="budgets"
     id=Column(Integer,primary_key=True,index=True)
-    amount=Column(Float,nullable=False)
+    amount=Column(Numeric(10,2),nullable=False)
     user_id=Column(Integer,ForeignKey("users.id"),nullable=False)
     month=Column(Integer,nullable=False)
     year=Column(Integer,nullable=False)
